+++ bciers/apps/reporting/src/app/components/finalReview/FinalReviewPage.tsx
@@ -1,6 +1,5 @@
 import { HasReportVersion } from "@reporting/src/app/utils/defaultPageFactoryTypes";
 import { getReportVerificationStatus } from "@reporting/src/app/utils/getReportVerificationStatus";
-import FinalReviewForm from "@reporting/src/app/components/finalReview/FinalReviewForm";
 import { getNavigationInformation } from "@reporting/src/app/components/taskList/navigationInformation";
 import { getOperationFacilitiesList } from "@reporting/src/app/utils/getOperationFacilitiesList";
 import {
@@ -9,7 +8,7 @@ import {
 } from "@reporting/src/app/components/taskList/types";
 import { getIsSupplementaryReport } from "@reporting/src/app/utils/getIsSupplementaryReport";
 import { getFinalReviewData } from "@reporting/src/app/utils/getFinalReviewData";
-import { FinalReviewFormNew } from "@reporting/src/app/components/finalReview/FinalReviewFormNew";
+import { FinalReviewForm } from "@reporting/src/app/components/finalReview/FinalReviewForm";
 
 export default async function FinalReviewPage({
   version_id,
@@ -44,18 +43,12 @@ export default async function FinalReviewPage({
 
   // If number of facilities is >40, we don't fetch the report data
   if (selectedFacilitiesCount > 40) {
-    return <FinalReviewForm navigationInformation={navInfo} data={[]} />;
+    return <FinalReviewForm navigationInformation={navInfo} data={null} />;
   }
 
-  // const flowData = await getFlowData(version_id);
   const finalReviewData = await getFinalReviewData(version_id);
-  // const finalReviewData = await reviewDataFactory(version_id, flowData);
 
   return (
-    <FinalReviewFormNew
-      navigationInformation={navInfo}
-      data={finalReviewData}
-    />
-    // <FinalReviewForm navigationInformation={navInfo} data={finalReviewData} />
+    <FinalReviewForm navigationInformation={navInfo} data={finalReviewData} />
   );
 }
